+++ apps/client/src/app2/App.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import { describe, it } from 'vitest';
-import App from '../app/App';
+import App from './App';
 
 describe('App rendered', () => {
   it('', () => {
