+++ apps/client/src/app2/App.tsx
@@ -1,5 +1,5 @@
-import { Header } from '../app/components/header/Header';
-import { Welcome } from '../app/pages/welcome/Welcome';
+import { Header } from '../app2/components/header/Header';
+import { Welcome } from './pages/welcome/Welcome';
 function App() {
   return (
     <div className="min-h-screen flex flex-col justify-center items-center bg-[url(/images/background-light.jpg)] dark:bg-[url(/images/background-dark.jpg)] bg-no-repeat bg-fixed bg-center bg-cover font-rub select-none text-black dark:text-white">
