+++ apps/client/src/app2/components/themeSwitch/ThemeSwitch.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, expect, it } from 'vitest';
-import { ThemeSwitch } from '../../../app/components/themeSwitch/ThemeSwitch';
+import { ThemeSwitch } from './ThemeSwitch';
 
 describe('ThemeSwitch', () => {
   it('change theme when you click on the switch', async () => {
