+++ apps/client/src/app2/pages/welcome/Welcome.tsx
